==> italy_venice/030-training_models/0.merge_files.ipynb <==
import os
import pandas as pd

path = '../020-clean_dbs/filtered/'

file_list = os.listdir(path)
columns_to_keep = ['osm_id', 'code', 'fclass', 'name', 'lon', 'lat', 'SELECT']
dfs = []

for filtered_df in file_list:
    df = pd.read_csv(path + filtered_df)
    dfs.append(df[columns_to_keep])

merged_df = pd.concat(dfs)
merged_df.to_csv('0. merged_df.csv')

merged_df
