==> italy_venice/030-training_models/3.exporting_tables.ipynb <==
import geopandas as gpd
from shapely.geometry import Polygon
import numpy as np

gdf = gpd.read_file('1.filtered_df.csv')
gdf

columns_to_normalize = ['nature', 'entartainment', 'transports', 'art', 'sport', 'tourism', 'security', 'accessibility']

for column in columns_to_normalize:
    print('Exporting: ', column)
    gdf[column] = [float(x) for x in gdf[column]]

    # normalization
    gdf[column] = (gdf[column] - gdf[column].mean())
    gdf[column] = gdf[column] / gdf[column].std()

    # removing outliers
    low = gdf[column].quantile(0.01)
    gdf.loc[gdf[column] < low, column] = low

    high = gdf[column].quantile(0.99)
    gdf.loc[gdf[column] > high, column] = high

gdf

lat_min_list = [float(x) for x in gdf.lat_min.to_list()]
lat_max_list = [float(x) for x in gdf.lat_max.to_list()]

lon_min_list = [float(x) for x in gdf.lon_min.to_list()]
lon_max_list = [float(x) for x in gdf.lon_max.to_list()]

polygons = []

for lat_min, lat_max, lon_min, lon_max in zip(lat_min_list, lat_max_list, lon_min_list, lon_max_list):
    polygons.append(Polygon([(lon_max, lat_max), (lon_min, lat_max), (lon_min, lat_min), (lon_max, lat_min)]))

gdf['geometry'] = polygons
gdf

gdf.to_csv('map_to_export.csv')
gdf.dtypes

exporting_df = gdf.drop(['geometry', 'row', 'column', 'lat_min', 'lat_max', 'lon_min', 'lon_max'], axis =1)
exporting_df['geometry'] = gdf.geometry
exporting_df

exporting_df.dtypes

exporting_df.to_csv('3.final_map_converted.csv')
