==> italy_venice/030-training_models/4.training_models.ipynb <==
import pandas as pd
import sklearn
import matplotlib as plt
from pycaret.regression import *

df = pd.read_csv('1.filtered_df.csv')
df.index = df['Unnamed: 0']
df.drop(['Unnamed: 0', 'row', 'column', 'lat_min', 'lat_max', 'lon_min', 'lon_max'], axis = 1, inplace = True)
df

df.describe()

exp = setup(df, target = 'tourism')

best = compare_models(n_select = 3)

evaluate_model(best[0])

dashboard(best[0])

tuned_model = tune_model(best[0])

# Expanding the Model - Interaction Effect and Polynomial Features

expanded_df = df.copy()
columns = expanded_df.columns.to_list()
columns.remove('tourism')

for a in columns:
    for b in columns:
        expanded_df[a + ' x ' + b ] = expanded_df[a] * expanded_df[b]

expanded_df

exp = setup(expanded_df, target = 'tourism')

best_expanded = compare_models()

evaluate_model(best_expanded)

# Ensemble Models

bagged = ensemble_model(best[0], method = 'Bagging')

boosting = ensemble_model(best[0], method = 'Boosting')

blender = blend_models(best)

stacking = stack_models(best)
